backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
python-dotenv
langchain-community
requests
python-dotenv
langgraph
langchain-google-community
